Project 1
This file was my first project that i made and it is just a simple code that lets you compare two spreadsheets together for buissness and compares the model number of each product in each spreadsheet and if its not in one, then the program will output a new spreadsheet and tell you which product is in which spreadsheet so you know if you sell it or if its a new item that you have to sell. This is mostly to help small businesses compare their sheets with manufacturers to cut down on labor costs.